Split buildability context helpers
- Move direct backend context and diagnostics into a dedicated helper

- Keep target capability checks free of backend-name string matching

- Tighten metrics budgets for the smaller buildability files
