+++ scripts/compiler-metrics.mts
@@ -25,7 +25,8 @@ const fileBudgets = {
   "native/zero-c/src/buildability.c": { maxLines: 260, maxStrcmpCalls: 2 },
   "native/zero-c/src/buildability.h": { maxLines: 20, maxStrcmpCalls: 0 },
   "native/zero-c/src/buildability_internal.h": { maxLines: 50, maxStrcmpCalls: 0 },
-  "native/zero-c/src/buildability_targets.c": { maxLines: 340, maxStrcmpCalls: 16 },
+  "native/zero-c/src/buildability_context.c": { maxLines: 190, maxStrcmpCalls: 16 },
+  "native/zero-c/src/buildability_targets.c": { maxLines: 155, maxStrcmpCalls: 0 },
   "native/zero-c/src/call_resolve.c": { maxLines: 200, maxStrcmpCalls: 2 },
   "native/zero-c/src/call_resolve.h": { maxLines: 100, maxStrcmpCalls: 0 },
   "native/zero-c/src/coff_format.c": { maxLines: 370, maxStrcmpCalls: 0 },
